create App components
