+++ src/components/Footer/Footer.js
@@ -1,0 +1,10 @@
+import FooterElement from './elements/FooterElement/FooterElement'
+import H2 from './elements/H2/H2'
+
+const Footer = () => {
+    FooterElement.appendChild(H2)
+
+    return FooterElement
+}
+
+export default Footer
