+++ src/components/Header/Header.js
@@ -1,5 +1,5 @@
-import HeaderElement from '../elements/HeaderElement/HeaderElement'
-import H1 from '../elements/H1/H1'
+import HeaderElement from './elements/HeaderElement/HeaderElement'
+import H1 from './elements/H1/H1'
 
 const Header = () => {
     HeaderElement.appendChild(H1)
